Chore and Drafts
I've brainstormed a wide range of data structures and algorithms in this commit. There may be errors or a lack of cohesive explanations, as it was done in a single sitting. I’m focusing on writing as much information and code as possible for now, with plans to go back later to refactor and improve the explanations.